refactor: remove providers bridge; refactor plugins bridging code
